+++ docs/changelog.md
@@ -175,3 +175,13 @@
 - Added code that resizes the backgrounds so they work correctly.
 - Implemented Map Select Screen so it is now functional.
 - Implemented Speed Increase mechanics, snake speed increases with each fruit consumed.
+- Added leaderboard page, method "setLeaderboardScene".
+- Implemented setLeaderboardScene.
+- Add CSS for leaderboard entries.
+- Added code for a text box to enter a username.
+- Added a button in the end game that saves the entry of the text box into a txt file and saves the score next to it.
+- Updated View class so that line length is shorter than 80 characters.
+- Added styling to the text field and button.
+- Implementing functionality that reads the text file, parses the output and adds it to the scoreboard.
+- Add a border to the scoreboard.
+- Move the header into a better position.